apps/bot/requirements.txt
# Discord bot requirements
discord.py==2.4.0
httpx==0.28.1
pydantic-settings==2.8.1
